eliminated disputed changes for now
SCCS-file: zdump.c
SCCS-SID: 7.70

+++ zdump.c
@@ -209,7 +209,7 @@ const char * const	zone;
 		return;
 	cp = abbrp;
 	wp = NULL;
-	while (isascii((int) *cp) && isalpha((int) *cp))
+	while (isascii(*cp) && isalpha(*cp))
 		++cp;
 	if (cp - abbrp == 0)
 		wp = _("lacks alphabetic at start");
@@ -219,7 +219,7 @@ const char * const	zone;
 		wp = _("has more than 6 alphabetics");
 	if (wp == NULL && (*cp == '+' || *cp == '-')) {
 		++cp;
-		if (isascii((int) *cp) && isdigit((int) *cp))
+		if (isascii(*cp) && isdigit(*cp))
 			if (*cp++ == '1' && *cp >= '0' && *cp <= '4')
 				++cp;
 	}
